Fixed Postmates missing storeUuid issue
The missing storeUuid issue has been fixed, at this point the error only seems to
be cause by that specific attribute, so I only accounted for that to avoid too many
checks that can result in performace issues when iterating long lists. If in the
future we see other issues with other attributes, we can add those as well. I tested
the fix with a know problamatic location and serach term.