Added Store information detail API endpoint
This PR adds store information for Grubhub and repurpouses Postamte's
getMenu method to get store information. The `/get` API was removed
in favor of using `/detail/store` as this will have both the store
data and the menu information for a store.

Additionally, as token data is being reused for Gruhub, a new method has
been added that will also need to be added to ther Services. This is the
`getValidToken()` method, which checks that the token being pulled from
the database using `getToken()` is always valid. If a token is not valid,
then `getValidToken()` will take the appropriate action to return a valid
token.