Search Data Processing and Reformating
Made "DataProcessing.mjs" file and "searchDataRemoveDuplicate" function that removes the duplicates at the highest precision and reformats the data as we decided. Integrated the function in the "search.mjs" return.

NOTE: To test the functionality you have to comment out or remove the stuff related to Doordash from "search()" in search.mjs